+++ packages/default-field-editors/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.67](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.7.66...@contentful/default-field-editors@1.7.67) (2025-06-03)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
 ## [1.7.66](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.7.65...@contentful/default-field-editors@1.7.66) (2025-05-26)
 
 **Note:** Version bump only for package @contentful/default-field-editors